+++ simple-server.py
@@ -7,7 +7,7 @@ import http.server
 import socketserver
 import os
 
-PORT = 8888
+PORT = 8080
 DIR = os.path.dirname(os.path.abspath(__file__))
 
 os.chdir(DIR)